requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
itsdangerous==2.2.0
